# src/main/resources/swagger_config/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8081
  description: Generated server url
paths:
...
  /api/v1/employee:
    get:
      tags:
      - Employee
      operationId: getAllEmployee
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommonResponseString'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/inline_response_400'
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseListEmployeeResponse'
    put:
      tags:
      - Employee
      operationId: updateEmployee
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEmployeeRequest'
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommonResponseString'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/inline_response_400'
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseEmployeeResponse'
...
  /api/v1/employee/{id}:
    get:
      tags:
      - Employee
      operationId: getEmployeeById
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommonResponseString'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/inline_response_400'
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseEmployeeResponse'
  /api/v1/employee/delete-account/{id}:
    delete:
      tags:
      - Employee
      operationId: deleteEmployeeById
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommonResponseString'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/inline_response_400'
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseString'
components:
  schemas:
    CommonResponseString:
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: string
    CommonResponseErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: '#/components/schemas/ErrorResponse'
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        status:
          type: integer
          format: int32
        timestamp:
          type: integer
          format: int64
...
    EmployeeResponse:
      type: object
      properties:
        employeeId:
          type: string
        employeeName:
          type: string
        department:
          type: string
...
    UpdateEmployeeRequest:
      type: object
      properties:
        employeeId:
          type: string
        employeeName:
          type: string
        department:
          type: string
        phoneNumber:
          type: string
        corporateEmail:
          type: string
    CommonResponseEmployeeResponse:
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: '#/components/schemas/EmployeeResponse'
...
    CommonResponseListEmployeeResponse:
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/EmployeeResponse'
    inline_response_400:
      oneOf:
      - $ref: '#/components/schemas/CommonResponseErrorResponse'
      - $ref: '#/components/schemas/CommonResponseString'